+++ admin/layouts/linker/details_above.php
@@ -32,17 +32,19 @@ $fields_tab_layout = 'fields_' . $layout_path_array[1];
 
 // get the fields
 $fields = $displayData->get($fields_tab_layout) ?: array(
-	'public_tagged_verses'
+	'name'
 );
 
 $hiddenFields = $displayData->get('hidden_fields') ?: array();
 
 ?>
 <?php if ($fields && count((array) $fields)) :?>
-<?php foreach($fields as $field): ?>
-	<?php if (in_array($field, $hiddenFields)) : ?>
-		<?php $form->setFieldAttribute($field, 'type', 'hidden'); ?>
-	<?php endif; ?>
-	<?php echo $form->renderField($field, null, null, array('class' => 'control-wrapper-' . $field)); ?>
-<?php endforeach; ?>
+<div class="form-inline form-inline-header">
+	<?php foreach($fields as $field): ?>
+		<?php if (in_array($field, $hiddenFields)) : ?>
+			<?php $form->setFieldAttribute($field, 'type', 'hidden'); ?>
+		<?php endif; ?>
+		<?php echo $form->renderField($field, null, null, array('class' => 'control-wrapper-' . $field)); ?>
+	<?php endforeach; ?>
+</div>
 <?php endif; ?>