+++ admin/views/linkers/tmpl/default_foot.php
@@ -20,5 +20,5 @@ defined('_JEXEC') or die('Restricted access');
 
 ?>
 <tr>
-	<td colspan="6"><?php echo $this->pagination->getListFooter(); ?></td>
+	<td colspan="8"><?php echo $this->pagination->getListFooter(); ?></td>
 </tr>